correct bugs in iks_ and iRs local index list, correct bugs in overload bcast functions, correct indices when calculating S(k), barrier_iR instead of iR
